requirements.txt
twitter>=1.17,<2
Mastodon.py>=1.0.8,<2
